댓글 기능 memberId 사용 -> JWT 사용으로 수정 (#150)
* feat: 댓글 기능에서 memberId 삭제 및 JWT로 로그인한 사용자 이메일 기반으로 멤버 조회하도록 변경

* refactor: getCurrentMember 메서드 SecurityUtils 클래스로 분리

* feat: 댓글 삭제 및 수정 요청 시 본인이 작성한 댓글인지 검증 구현

* refactor: CommentCreateRequest 내 불필요한 memberId 필드 제거

* test: 댓글 기능 변경점에 따라 테스트 코드 수정 및 전체 통과 완료

* test: 댓글 수정, 삭제 시 권한 없음 예외 테스트코드 작성 및 통과

* feat: SecurityUtils 클래스 기본 생성자 접근 방지 구현

+++ src/main/java/balancetalk/module/comment/application/CommentService.java
@@ -22,9 +22,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+
+import static balancetalk.global.exception.ErrorCode.*;
+import static balancetalk.global.utils.SecurityUtils.getCurrentMember;
 
 @Service
 @Transactional
@@ -41,10 +45,10 @@ public class CommentService {
     private int maxDepth;
 
     public Comment createComment(CommentRequest request, Long postId) {
-        Member member = validateMemberId(request);
+        Member member = getCurrentMember(memberRepository);
         Post post = validatePostId(postId);
         BalanceOption balanceOption = validateBalanceOptionId(request, post);
-        voteRepository.findByMemberIdAndBalanceOption_PostId(request.getMemberId(), postId)
+        voteRepository.findByMemberIdAndBalanceOption_PostId(member.getId(), postId)
                 .orElseThrow(() -> new BalanceTalkException(ErrorCode.NOT_FOUND_VOTE));
 
         Comment comment = request.toEntity(member, post);
@@ -52,7 +56,7 @@ public class CommentService {
     }
 
     @Transactional(readOnly = true)
-    public List<CommentResponse> findAll(Long postId) { // TODO: 탈퇴한 회원의 정보는 어떻게 표시되는가?
+    public List<CommentResponse> findAll(Long postId) {
         validatePostId(postId);
 
         List<Comment> comments = commentRepository.findByPostId(postId);
@@ -72,12 +76,22 @@ public class CommentService {
     public Comment updateComment(Long commentId, String content) {
         Comment comment = validateCommentId(commentId);
 
+        if (!getCurrentMember(memberRepository).equals(comment.getMember())) {
+            throw new BalanceTalkException(FORBIDDEN_COMMENT_MODIFY);
+        }
+
         comment.updateContent(content);
         return comment;
     }
 
     public void deleteComment(Long commentId) {
         validateCommentId(commentId);
+        Member commentMember = commentRepository.findById(commentId).get().getMember();
+
+        if (!getCurrentMember(memberRepository).equals(commentMember)) {
+            throw new BalanceTalkException(FORBIDDEN_COMMENT_DELETE);
+        }
+
         commentRepository.deleteById(commentId);
     }
 
@@ -89,11 +103,7 @@ public class CommentService {
         Comment parentComment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_COMMENT));
 
-        Member member = memberRepository.findById(request.getMemberId())
-                .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_MEMBER));
-
-
-
+        Member member = getCurrentMember(memberRepository);
 
         // 부모 댓글과 연결된 게시글이 맞는지 확인
         if (!parentComment.getPost().equals(post)) {
@@ -104,11 +114,6 @@ public class CommentService {
 
         Comment reply = request.toEntity(member, post, parentComment);
         return commentRepository.save(reply);
-    }
-
-    private Member validateMemberId(CommentRequest request) { // TODO: validate 메서드 분리 재고
-        return memberRepository.findById(request.getMemberId())
-                .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_MEMBER));
     }
 
     private Post validatePostId(Long postId) {
@@ -143,12 +148,11 @@ public class CommentService {
         }
         return depth;
     }
-      
-    public Long likeComment(Long postId, Long commentId, Long memberId) {
+
+    public Long likeComment(Long postId, Long commentId) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_COMMENT));
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new BalanceTalkException(ErrorCode.NOT_FOUND_MEMBER));
+        Member member = getCurrentMember(memberRepository);
 
         if (commentLikeRepository.existsByMemberAndComment(member, comment)) {
             throw new BalanceTalkException(ErrorCode.ALREADY_LIKE_COMMENT);
@@ -163,11 +167,10 @@ public class CommentService {
         return comment.getId();
     }
 
-    public void cancelLikeComment(Long commentId, Long memberId) {
+    public void cancelLikeComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_COMMENT));
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new BalanceTalkException(ErrorCode.NOT_FOUND_MEMBER));
+        Member member = getCurrentMember(memberRepository);
 
         commentLikeRepository.deleteByMemberAndComment(member, comment);
     }
